Added files via upload
I just doodled this real quick in Inkscape, and it is an idea for either an antagonist or the main protagonist of the game. This is just a front facing character, no side views, back views, or animations yet since I wanted to get input before I did anything else.

Tell me what you guys think, and I will edit the image accordingly, scrap it, or even spend hours upon hours trying to make a highly detailed ninja image that we could use within the game.

PS: I was thinking of maybe doing a wizard character sprite along with a biped monster with long spear like talons that it balances on like Chris described in our meeting. The second one could be easily scaled down and had its damage reduced to be a regular enemy that is encountered. Also, Chris, if you have any art stuff finished that you plan on using for the project that Jon and you have both already agreed on, could you upload them so I can see a general art style that we plan on going for in this project? 

Thanks.